+++ packages/tracing/src/integrations/index.ts
@@ -1,4 +1,4 @@
-export { Express } from './express';
-export { Postgres } from './postgres';
-export { Mysql } from './mysql';
-export { Mongo } from './mongo';
+export { Express } from './node/express';
+export { Postgres } from './node/postgres';
+export { Mysql } from './node/mysql';
+export { Mongo } from './node/mongo';
