+++ packages/tracing/test/integrations/mongo.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/unbound-method */
 import { Hub, Scope } from '@sentry/hub';
 
-import { Mongo } from '../../src/integrations/mongo';
+import { Mongo } from '../../src/integrations/node/mongo';
 import { Span } from '../../src/span';
 
 class Collection {
